+++ probeye/inference/scipy/solver.py
@@ -29,7 +29,10 @@ class ScipySolver(Solver):
     """
 
     def __init__(
-        self, problem: "InverseProblem", seed: int = 1, show_progress: bool = True
+        self,
+        problem: "InverseProblem",
+        seed: Optional[int] = None,
+        show_progress: bool = True,
     ):
         logger.debug("Initializing ScipySolver")
         super().__init__(problem, seed=seed, show_progress=show_progress)
@@ -77,12 +80,14 @@ class ScipySolver(Solver):
         Translate the inverse problem's forward models as needed for this solver.
         """
         logger.debug("Translate the problem's forward models")
-        for fwd_name in self.problem.forward_models:
+        for fwd_name, fwd_obj in self.problem.forward_models.items():
 
             # create a full forward model from its hull where the sensors are not yet
             # connected to the experimental data
             forward_model_hull = self.problem.forward_models[fwd_name]
-            forward_model = forward_model_hull.__class__.__bases__[0](fwd_name)
+            forward_model = forward_model_hull.__class__.__bases__[0](
+                fwd_name, *fwd_obj.args, **fwd_obj.kwargs
+            )
             synchronize_objects(
                 forward_model,
                 forward_model_hull,
@@ -485,7 +490,10 @@ class MaxLikelihoodSolver(ScipySolver):
     """
 
     def __init__(
-        self, problem: "InverseProblem", seed: int = 1, show_progress: bool = True
+        self,
+        problem: "InverseProblem",
+        seed: Optional[int] = None,
+        show_progress: bool = True,
     ):
         logger.debug(f"Initializing {self.__class__.__name__}")
         super().__init__(problem, seed=seed, show_progress=show_progress)
@@ -522,7 +530,10 @@ class MaxPosteriorSolver(ScipySolver):
     """
 
     def __init__(
-        self, problem: "InverseProblem", seed: int = 1, show_progress: bool = True
+        self,
+        problem: "InverseProblem",
+        seed: Optional[int] = None,
+        show_progress: bool = True,
     ):
         logger.debug(f"Initializing {self.__class__.__name__}")
         super().__init__(problem, seed=seed, show_progress=show_progress)